--- text/old_wsgi.py
@@ -1,4 +1,0 @@
-from web import app
-  
-if __name__ == "__main__":
-        app.run(debug=True, host='0.0.0.0', port='5010')
